Show the network buffer if it is selected by any BufferView.
Fixes a bug where selecting the a network buffer after the last visible
buffer had been deleted by always showing a network buffer if it is
selected. This still leaves the problem of network buffers not being
properly filtered out. This is probably due to the selection model not
triggering to the filter model.